Create placeholder functions and fix circular import
