+++ tests/variable_definitions/test_vardef.py
@@ -21,15 +21,6 @@ from dapla_metadata.variable_definitions.generated.vardef_client.configuration i
 from dapla_metadata.variable_definitions.generated.vardef_client.models.draft import (
     Draft,
 )
-from dapla_metadata.variable_definitions.generated.vardef_client.models.patch import (
-    Patch,
-)
-from dapla_metadata.variable_definitions.generated.vardef_client.models.update_draft import (
-    UpdateDraft,
-)
-from dapla_metadata.variable_definitions.generated.vardef_client.models.validity_period import (
-    ValidityPeriod,
-)
 from dapla_metadata.variable_definitions.generated.vardef_client.models.variable_status import (
     VariableStatus,
 )
@@ -141,30 +132,6 @@ def test_not_found(
     assert str(e.value) == "Status 404: Not found"
 
 
-def test_list_patches(client_configuration: Configuration):
-    VardefClient.set_config(client_configuration)
-    landbak = Vardef.get_variable_definition_by_id(
-        variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
-    )
-    assert isinstance(landbak.list_patches()[0], CompletePatchOutput)
-
-
-def test_get_patch(client_configuration: Configuration):
-    VardefClient.set_config(client_configuration)
-    landbak = Vardef.get_variable_definition_by_id(
-        variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
-    )
-    assert isinstance(landbak.get_patch(1), CompletePatchOutput)
-
-
-def test_list_validity_periods(client_configuration: Configuration):
-    VardefClient.set_config(client_configuration)
-    landbak = Vardef.get_variable_definition_by_id(
-        variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
-    )
-    assert isinstance(landbak.list_validity_periods()[0], CompletePatchOutput)
-
-
 def test_create_draft(
     monkeypatch: pytest.MonkeyPatch,
     client_configuration: Configuration,
@@ -194,69 +161,6 @@ def test_migrate_from_vardok(
     assert my_draft.id is not None
     assert my_draft.patch_id == 1
     assert my_draft.variable_status == VariableStatus.DRAFT
-
-
-def test_update_draft(
-    monkeypatch: pytest.MonkeyPatch,
-    client_configuration: Configuration,
-    update_draft: UpdateDraft,
-):
-    monkeypatch.setenv(DAPLA_GROUP_CONTEXT, VARDEF_EXAMPLE_ACTIVE_GROUP)
-    VardefClient.set_config(client_configuration)
-    my_draft = Vardef.get_variable_definition_by_id(
-        variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
-    )
-    assert isinstance(my_draft.update_draft(update_draft), CompletePatchOutput)
-
-
-def test_delete_draft(
-    monkeypatch: pytest.MonkeyPatch,
-    client_configuration: Configuration,
-):
-    monkeypatch.setenv(DAPLA_GROUP_CONTEXT, VARDEF_EXAMPLE_ACTIVE_GROUP)
-    VardefClient.set_config(client_configuration)
-    my_draft = Vardef.get_variable_definition_by_id(
-        variable_definition_id=VARDEF_EXAMPLE_DEFINITION_ID,
-    )
-    assert my_draft.id is not None
-    result = my_draft.delete_draft()
-    assert result == f"Variable {VARDEF_EXAMPLE_DEFINITION_ID} safely deleted"
-
-
-def test_create_patch(
-    monkeypatch: pytest.MonkeyPatch,
-    client_configuration: Configuration,
-    patch: Patch,
-    variable_definition: VariableDefinition,
-):
-    monkeypatch.setenv(DAPLA_GROUP_CONTEXT, VARDEF_EXAMPLE_ACTIVE_GROUP)
-    VardefClient.set_config(client_configuration)
-    my_variable = variable_definition
-    assert my_variable.patch_id == 1
-    created_patch = my_variable.create_patch(
-        patch=patch,
-        valid_from=VARDEF_EXAMPLE_DATE,
-    )
-    assert isinstance(
-        created_patch,
-        CompletePatchOutput,
-    )
-    assert created_patch.patch_id == PATCH_ID
-
-
-def test_create_validity_period(
-    monkeypatch: pytest.MonkeyPatch,
-    client_configuration: Configuration,
-    validity_period: ValidityPeriod,
-    variable_definition: VariableDefinition,
-):
-    monkeypatch.setenv(DAPLA_GROUP_CONTEXT, VARDEF_EXAMPLE_ACTIVE_GROUP)
-    VardefClient.set_config(client_configuration)
-    my_variable = variable_definition
-    assert isinstance(
-        my_variable.create_validity_period(validity_period),
-        CompletePatchOutput,
-    )
 
 
 @pytest.mark.usefixtures("set_temp_workspace")
